requirements.txt
requests
python-dotenv
tweepy
asyncio
openai
psycopg2-binary
httpx
sqlalchemy